Add files via upload
The project is a comprehensive analysis that utilizes Python, Excel, and Tableau to delve into the patterns of bike-sharing usage in London. With an interactive Tableau dashboard, it showcases trends, user behaviors, and environmental impacts, while also forecasting future demand. Based on a dataset from Kaggle, the project reveals key insights on the optimal days for bike-sharing, aiming to improve both service planning and user experience.